+++ src/lando/api/tests/conftest.py
@@ -1,8 +1,11 @@
 import json
 import os
+import pathlib
+from collections.abc import Callable
 from datetime import datetime, timezone
 from pathlib import Path
 
+import py
 import pytest
 import redis
 import requests_mock
@@ -618,3 +621,103 @@ def proxy_client(monkeypatch, fake_request):
 def authenticated_client(user, user_plaintext_password, client):
     client.login(username=user.username, password=user_plaintext_password)
     return client
+
+
+@pytest.mark.django_db
+def hg_repo_mc(
+    hg_server: str,
+    hg_clone: py.path,
+    *,
+    approval_required: bool = False,
+    autoformat_enabled: bool = False,
+    force_push: bool = False,
+    push_target: str = "",
+) -> Repo:
+    params = {
+        "required_permission": SCM_LEVEL_3,
+        "url": hg_server,
+        "push_path": hg_server,
+        "pull_path": hg_server,
+        "system_path": hg_clone.strpath,
+        # The option below can be overriden in the parameters
+        "approval_required": approval_required,
+        "autoformat_enabled": autoformat_enabled,
+        "force_push": force_push,
+        "push_target": push_target,
+    }
+    repo = Repo.objects.create(
+        scm_type=SCM_TYPE_HG,
+        name="mozilla-central-hg",
+        **params,
+    )
+    repo.save()
+    return repo
+
+
+@pytest.mark.django_db
+def git_repo_mc(
+    git_repo: pathlib.Path,
+    tmp_path: pathlib.Path,
+    *,
+    approval_required: bool = False,
+    autoformat_enabled: bool = False,
+    force_push: bool = False,
+    push_target: str = "",
+) -> Repo:
+    repos_dir = tmp_path / "repos"
+    repos_dir.mkdir()
+
+    params = {
+        "required_permission": SCM_LEVEL_3,
+        "url": str(git_repo),
+        "push_path": str(git_repo),
+        "pull_path": str(git_repo),
+        "system_path": repos_dir / "git_repo",
+        # The option below can be overriden in the parameters
+        "approval_required": approval_required,
+        "autoformat_enabled": autoformat_enabled,
+        "force_push": force_push,
+        "push_target": push_target,
+    }
+
+    repo = Repo.objects.create(
+        scm_type=SCM_TYPE_GIT,
+        name="mozilla-central-git",
+        **params,
+    )
+    repo.save()
+    repo.scm.prepare_repo(repo.pull_path)
+    return repo
+
+
+@pytest.fixture()
+def repo_mc(
+    # Git
+    git_repo: pathlib.Path,
+    tmp_path: pathlib.Path,
+    # Hg
+    hg_server: str,
+    hg_clone: py.path,
+) -> Callable:
+    def factory(
+        scm_type: str,
+        *,
+        approval_required: bool = False,
+        autoformat_enabled: bool = False,
+        force_push: bool = False,
+        push_target: str = "",
+    ) -> Repo:
+        params = {
+            "approval_required": approval_required,
+            "autoformat_enabled": autoformat_enabled,
+            "force_push": force_push,
+            "push_target": push_target,
+        }
+
+        if scm_type == SCM_TYPE_GIT:
+            return git_repo_mc(git_repo, tmp_path, **params)
+        elif scm_type == SCM_TYPE_HG:
+            return hg_repo_mc(hg_server, hg_clone, **params)
+        raise Exception(f"Unknown SCM Type {scm_type=}")
+
+    return factory
