+++ src/lando/api/tests/test_landings.py
@@ -1,13 +1,9 @@
 import io
-import pathlib
 import unittest.mock as mock
-from collections.abc import Callable
 
-import py
 import pytest
 
 from lando.api.legacy.workers.landing_worker import AUTOFORMAT_COMMIT_MESSAGE
-from lando.main.models import SCM_LEVEL_3, Repo
 from lando.main.models.landing_job import (
     LandingJob,
     LandingJobStatus,
@@ -232,106 +228,6 @@ TeSt
 aDdInG AnOtHeR LiNe
 aDd oNe mOrE LiNe
 """.lstrip()
-
-
-@pytest.mark.django_db
-def hg_repo_mc(
-    hg_server: str,
-    hg_clone: py.path,
-    *,
-    approval_required: bool = False,
-    autoformat_enabled: bool = False,
-    force_push: bool = False,
-    push_target: str = "",
-) -> Repo:
-    params = {
-        "required_permission": SCM_LEVEL_3,
-        "url": hg_server,
-        "push_path": hg_server,
-        "pull_path": hg_server,
-        "system_path": hg_clone.strpath,
-        # The option below can be overriden in the parameters
-        "approval_required": approval_required,
-        "autoformat_enabled": autoformat_enabled,
-        "force_push": force_push,
-        "push_target": push_target,
-    }
-    repo = Repo.objects.create(
-        scm_type=SCM_TYPE_HG,
-        name="mozilla-central-hg",
-        **params,
-    )
-    repo.save()
-    return repo
-
-
-@pytest.mark.django_db
-def git_repo_mc(
-    git_repo: pathlib.Path,
-    tmp_path: pathlib.Path,
-    *,
-    approval_required: bool = False,
-    autoformat_enabled: bool = False,
-    force_push: bool = False,
-    push_target: str = "",
-) -> Repo:
-    repos_dir = tmp_path / "repos"
-    repos_dir.mkdir()
-
-    params = {
-        "required_permission": SCM_LEVEL_3,
-        "url": str(git_repo),
-        "push_path": str(git_repo),
-        "pull_path": str(git_repo),
-        "system_path": repos_dir / "git_repo",
-        # The option below can be overriden in the parameters
-        "approval_required": approval_required,
-        "autoformat_enabled": autoformat_enabled,
-        "force_push": force_push,
-        "push_target": push_target,
-    }
-
-    repo = Repo.objects.create(
-        scm_type=SCM_TYPE_GIT,
-        name="mozilla-central-git",
-        **params,
-    )
-    repo.save()
-    repo.scm.prepare_repo(repo.pull_path)
-    return repo
-
-
-@pytest.fixture()
-def repo_mc(
-    # Git
-    git_repo: pathlib.Path,
-    tmp_path: pathlib.Path,
-    # Hg
-    hg_server: str,
-    hg_clone: py.path,
-) -> Callable:
-    def factory(
-        scm_type: str,
-        *,
-        approval_required: bool = False,
-        autoformat_enabled: bool = False,
-        force_push: bool = False,
-        push_target: str = "",
-    ) -> Repo:
-        params = {
-            "approval_required": approval_required,
-            "autoformat_enabled": autoformat_enabled,
-            "force_push": force_push,
-            "push_target": push_target,
-        }
-
-        if scm_type == SCM_TYPE_GIT:
-            return git_repo_mc(git_repo, tmp_path, **params)
-        elif scm_type == SCM_TYPE_HG:
-            return hg_repo_mc(hg_server, hg_clone, **params)
-        raise Exception(f"Unknown SCM Type {scm_type=}")
-
-    return factory
 
 
 @pytest.mark.parametrize(
